Change parser for repo2 commands Signed-off-by: Romain Keramitas  <r.keramitas@gmail.com>
