+++ sourced/ml/__main__.py
@@ -10,12 +10,11 @@ from sourced.ml.cmd_entries import bigartm2asdf_entry, dump_model, projector_ent
     run_swivel, postprocess_id2vec, preprocess_id2vec, repos2coocc_entry, repos2df_entry, \
     repos2ids_entry, repos2bow_entry, repos2roles_and_ids_entry, repos2id_distance_entry, \
     repos2id_sequence_entry
-from sourced.ml.cmd_entries.args import add_repo2_args, add_feature_args, \
-    add_vocabulary_size_arg, add_extractor_args, add_split_stem_arg, \
-    ArgumentDefaultsHelpFormatterNoNone
-from sourced.ml.cmd_entries.repos2bow import add_bow_args
+from sourced.ml.cmd_entries.args import add_df_args, add_feature_args, \
+    add_vocabulary_size_arg, add_repo2_args, add_split_stem_arg, \
+    ArgumentDefaultsHelpFormatterNoNone, add_bow_args
 from sourced.ml.cmd_entries.run_swivel import mirror_tf_args
-from sourced.ml.utils import install_bigartm, add_engine_args
+from sourced.ml.utils import install_bigartm
 
 
 def get_parser() -> argparse.ArgumentParser:
@@ -30,40 +29,34 @@ def get_parser() -> argparse.ArgumentParser:
     # Create and construct subparsers
     subparsers = parser.add_subparsers(help="Commands", dest="command")
 
-    def add_parser(name, help):
+    def add_parser(name, help_message):
         return subparsers.add_parser(
-            name, help=help, formatter_class=ArgumentDefaultsHelpFormatterNoNone)
+            name, help=help_message, formatter_class=ArgumentDefaultsHelpFormatterNoNone)
 
     # ------------------------------------------------------------------------
     repos2bow_parser = add_parser(
         "repos2bow", "Convert source code to the bag-of-words model.")
     repos2bow_parser.set_defaults(handler=repos2bow_entry)
+    add_df_args(repos2bow_parser)
     add_repo2_args(repos2bow_parser)
-    add_engine_args(repos2bow_parser)
-    add_bow_args(repos2bow_parser)
     add_feature_args(repos2bow_parser)
+    add_bow_args(repos2bow_parser)
     # ------------------------------------------------------------------------
     repos2df_parser = add_parser(
         "repos2df", "Calculate document frequencies of features extracted from source code.")
     repos2df_parser.set_defaults(handler=repos2df_entry)
+    add_df_args(repos2df_parser)
     add_repo2_args(repos2df_parser)
-    add_engine_args(repos2df_parser)
     add_feature_args(repos2df_parser)
     # ------------------------------------------------------------------------
     repos2ids_parser = subparsers.add_parser(
         "repos2ids", help="Convert source code to a bag of identifiers.")
     repos2ids_parser.set_defaults(handler=repos2ids_entry)
-    add_engine_args(repos2ids_parser)
-    repos2ids_parser.add_argument(
-        "-r", "--repositories", required=True,
-        help="The path to the repositories.")
+    add_repo2_args(repos2ids_parser)
+    add_split_stem_arg(repos2ids_parser)
     repos2ids_parser.add_argument(
         "-o", "--output", required=True,
         help="[OUT] output path to the CSV file with identifiers.")
-    repos2ids_parser.add_argument(
-        "--split", action="store_true",
-        help="Enables filtering identifiers that are splittable"
-             "based on special characters or case changes.")
     repos2ids_parser.add_argument(
         "--idfreq", action="store_true",
         help="Adds identifier frequencies to the output CSV file."
@@ -74,8 +67,8 @@ def get_parser() -> argparse.ArgumentParser:
     repos2coocc_parser = add_parser(
         "repos2coocc", "Convert source code to the sparse co-occurrence matrix of identifiers.")
     repos2coocc_parser.set_defaults(handler=repos2coocc_entry)
-    add_engine_args(repos2coocc_parser)
-    add_repo2_args(repos2coocc_parser, quant=False)
+    add_df_args(repos2coocc_parser)
+    add_repo2_args(repos2coocc_parser)
     add_split_stem_arg(repos2coocc_parser)
     repos2coocc_parser.add_argument(
         "-o", "--output", required=True,
@@ -85,8 +78,7 @@ def get_parser() -> argparse.ArgumentParser:
         "repos2roles_ids", "Converts a UAST to a list of pairs, where pair is a role and "
         "identifier. Role is merged generic roles where identifier was found.")
     repos2roles_and_ids.set_defaults(handler=repos2roles_and_ids_entry)
-    add_engine_args(repos2roles_and_ids)
-    add_extractor_args(repos2roles_and_ids)
+    add_repo2_args(repos2roles_and_ids)
     add_split_stem_arg(repos2roles_and_ids)
     repos2roles_and_ids.add_argument(
         "-o", "--output", required=True,
@@ -97,8 +89,7 @@ def get_parser() -> argparse.ArgumentParser:
         "repos2id_distance", "Converts a UAST to a list of identifier pairs "
                              "and distance between them.")
     repos2identifier_distance.set_defaults(handler=repos2id_distance_entry)
-    add_engine_args(repos2identifier_distance)
-    add_extractor_args(repos2identifier_distance)
+    add_repo2_args(repos2identifier_distance)
     add_split_stem_arg(repos2identifier_distance)
     repos2identifier_distance.add_argument(
         "-t", "--type", required=True, choices=IdentifierDistance.DistanceType.All,
@@ -115,8 +106,7 @@ def get_parser() -> argparse.ArgumentParser:
         "repos2id_sequence", "Converts a UAST to sequence of identifiers sorted by "
                              "order of appearance.")
     repos2id_sequence.set_defaults(handler=repos2id_sequence_entry)
-    add_engine_args(repos2id_sequence)
-    add_extractor_args(repos2id_sequence)
+    add_repo2_args(repos2id_sequence)
     add_split_stem_arg(repos2id_sequence)
     repos2id_sequence.add_argument(
         "--skip-docname", default=False, action="store_true",
