+++ sourced/ml/cmd_entries/repos2bow.py
@@ -1,4 +1,3 @@
-import argparse
 import logging
 from uuid import uuid4
 
@@ -9,16 +8,6 @@ from sourced.ml.transformers import Ignition, UastExtractor, UastDeserializer, U
     Indexer, UastRow2Document, BOWWriter, Moder, LanguageSelector, create_parquet_loader
 from sourced.ml.utils import create_engine
 from sourced.ml.utils.engine import pipeline_graph, pause
-
-
-def add_bow_args(my_parser: argparse.ArgumentParser):
-    my_parser.add_argument(
-        "--bow", required=True, help="[OUT] The path to the Bag-Of-Words model.")
-    my_parser.add_argument(
-        "--batch", default=BOWWriter.DEFAULT_CHUNK_SIZE, type=int,
-        help="The maximum size of a single BOW file in bytes.")
-    my_parser.add_argument(
-        "--parquet", action="store_true", help="If it's parquet input.")
 
 
 @pause
